+++ spec/actor_spec.rb
@@ -167,6 +167,46 @@ RSpec.describe Actor do
       end
     end
 
+    context "when using prompts" do
+      context "when using `call`" do
+        let(:result) { PromptForInput.call }
+
+        before { PromptForInput.prompt.input.truncate(0) }
+
+        after { PromptForInput.prompt.input.truncate(0) }
+
+        it "has answer available" do
+          expect(result.answer).to eq("YARR PFI")
+        end
+
+        it "can take user input" do
+          PromptForInput.prompt.input << "Hargle\n"
+          PromptForInput.prompt.input.rewind
+
+          expect(result.answer).to eq("Hargle")
+        end
+      end
+
+      context "when using `play`" do
+        let(:result) { PlayWithPromptForInput.call }
+
+        before { PlayWithPromptForInput.prompt.input.truncate(0) }
+
+        after { PlayWithPromptForInput.prompt.input.truncate(0) }
+
+        it "has answer available" do
+          expect(result.answer).to eq("YARR PWPFI")
+        end
+
+        it "can take user input" do
+          PlayWithPromptForInput.prompt.input << "Anthropomorphize\n"
+          PlayWithPromptForInput.prompt.input.rewind
+
+          expect(result.answer).to eq("Anthropomorphize")
+        end
+      end
+    end
+
     context "when playing several actors and one fails" do
       let(:result) { ServiceActor::Result.new(value: 0) }