+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "bundler/setup"
+
 require "service_actor"
 require "pry"